fix(cloud): Keep another meter's allowance intact
A meter this plugin does not count was reported as unlimited rather than
as untouched, so a plan that allowed a fixed number of anything else would
have been unenforceable.

A plan can now state an unlimited token allowance as null, the way its
limits already do. Leaving the allowance out still means none.

+++ config/cloudplans.php
@@ -75,7 +75,7 @@ return [
                 'max_wallets' => 3,
                 'max_categories' => 10,
             ],
-            'token_allowance' => (int) env('CLOUD_PLAN_FREE_TOKEN_ALLOWANCE', 0),
+            'token_allowance' => env('CLOUD_PLAN_FREE_TOKEN_ALLOWANCE', 0),
             // The strings below are marketing copy. What the gate reads is
             // feature_keys, limits and token_allowance above.
             'features' => [
@@ -97,7 +97,7 @@ return [
                 'max_wallets' => null,
                 'max_categories' => null,
             ],
-            'token_allowance' => (int) env('CLOUD_PLAN_MONTHLY_TOKEN_ALLOWANCE', 50000),
+            'token_allowance' => env('CLOUD_PLAN_MONTHLY_TOKEN_ALLOWANCE', 50000),
             'features' => [
                 'Unlimited categories and wallets',
                 'Mobile and web access',
@@ -118,7 +118,7 @@ return [
                 'max_wallets' => null,
                 'max_categories' => null,
             ],
-            'token_allowance' => (int) env('CLOUD_PLAN_YEARLY_TOKEN_ALLOWANCE', 50000),
+            'token_allowance' => env('CLOUD_PLAN_YEARLY_TOKEN_ALLOWANCE', 50000),
             'features' => [
                 'Everything in Monthly',
                 '2 months free (save ~17%)',
